hugo/fogros2: Launch rosbridge and foxglove
Use the python subprocess module to ssh into the ec2 instance and launch rosbridge to solve issues with paramiko and background processes on the cloud. Use Eric's code for starting docker containers to start the foxglove docker. Users can view foxglove on port 8080 of theec2's ip address and connect to rosbridge on port 9090 of the same ip address.